LayoutTests/fast/mutation/observe-subtree-expected.txt

Test WebKitMutationObserver.observe on a subtree

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing basic aspects of subtree observation.
...attribute and characterData changes in subtree
PASS mutations.length is 2
PASS mutations[0].type is "attributes"
PASS mutations[0].target is subDiv
PASS mutations[0].attributeName is "foo"
PASS mutations[0].attributeNamespace is null
PASS mutations[1].type is "characterData"
PASS mutations[1].target is subDiv.firstChild

Testing two observers at different depths.
PASS mutations.length is 1
PASS mutations[0].type is "attributes"
PASS mutations[0].target is subDiv
PASS mutations[0].attributeName is "foo"
PASS mutations[0].attributeNamespace is null
PASS mutations2.length is 1
PASS mutations2[0].type is "attributes"
PASS mutations2[0].target is subDiv
PASS mutations2[0].attributeName is "foo"
PASS mutations2[0].attributeNamespace is null

Testing one observer at two different depths.
PASS calls is 1
PASS mutations.length is 1
PASS mutations[0].type is "attributes"
PASS mutations[0].target is subDiv
PASS mutations[0].attributeName is "foo"
PASS mutations[0].attributeNamespace is null

PASS successfullyParsed is true

TEST COMPLETE
